refactor: ui 폴더를 components 폴더 내 common으로 변경
