+++ frontend/src/components/common/Button/ButtonLogin.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../styles/LoginSignup.css"
+import "../../../styles/LoginSignup.css"
 
 const Button = ({ text, onClick, disabled }) => {
   return (
