+++ frontend/src/pages/SignupPage.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import Header from "../components/Header";
 import LogInContainer from "../components/LogInContainer";
 import InputField from "../components/InputField";
-import Button from "../ui/Button/ButtonLogin";
+import Button from "../components/common/Button/ButtonLogin";
 import "../styles/LoginSignup.css"
 
 const SignupPage = () => {
